requirements.txt
langchain==0.3.25
langchain-openai==0.3.12
openai>=1.0.0,<2.0.0
python-dotenv
flask
